Create Special number or not
Write a program to check wheather the given number is special number or not.
What is special number?
Split the number to digits. Add the digits, multiply the digit, add the output of addition and multiplication and the
resultant number should be same as the original number.

Example: If the input is 19 ====>>1+9= 10 and 1*9 ====>> 10+9=19 which is equal to the input so the given number
is a special number.

Number should be a 2 digit number. Otherwise print "Invalid input"



'''
TEST CASE 1:
INPUT:
3
19
11
5


OUTPUT:
Special number
Not a special number
Invalid input