Create Great War (count number of 1's)
Great war is on - Each and every living being is giving their best to save the world from the dead. Arya fighting with full
dedication found that daemons need  to be killed with a pattern or else they wont die. so that she can fight till the end of 
the war. she tried to undestand the pattern. she kills 'Xth' daemon by stabbing them 'N' times.

Daemon approaching order (X) Number of times-stabbing(N)

1           1
2           1
3           2
.           .

55           5
.           .

98           3
101           4
198           4

This is the hint of the pattern that Arya needs to follow. Help Arya!


TEST CASE 1:
INPUT:
4
102
95
72
60


OUTPUT:
4
6
2
4